Revert "Revert "Merge branch 'master' of https://github.com/Eytins/HotelManagement""
This reverts commit 5011b3710f56d9ce33f1e774cfccac658d36a519.

+++ src/com/HotelManagement/dao/hotelType/HotelTypeMapper.java
@@ -33,7 +33,7 @@ public interface HotelTypeMapper {
 
     //更改房型
     int modifyRoomType(@Param("roomTypeId") Integer roomTypeId,
-                       @Param("price") Integer price,
+                       @Param("price") long price,
                        @Param("numberOfPerson") Integer numberOfPerson);
